# day02-tlias-web-management/Web开发-2024.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web开发-2024",
    "description": "新版本的Web开发的接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/students": {
      "get": {
        "summary": "分页查询",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "学生信息的分页条件查询",
        "operationId": "pageUsingGET_2",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "clazzId",
            "in": "query",
            "description": "班级ID",
            "required": false,
            "example": "3",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "degree",
            "in": "query",
            "description": "学历",
            "required": false,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "姓名",
            "required": false,
            "example": "张三",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": false,
            "example": "10",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {},
                  "x-apifox-orders": [
                    "code",
                    "msg",
                    "data"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "type": "object",
                      "x-apifox-refs": {},
                      "x-apifox-orders": [
                        "total",
                        "rows"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer",
                          "title": "总记录数",
                          "description": "总记录数"
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "format": "int32",
                                "title": "ID",
                                "additionalProperties": false
                              },
                              "name": {
                                "type": "string",
                                "title": "姓名",
                                "additionalProperties": false
                              },
                              "no": {
                                "type": "string",
                                "title": "学号",
                                "additionalProperties": false
                              },
                              "gender": {
                                "type": "integer",
                                "format": "int32",
                                "title": "性别",
                                "description": "1: 男 , 2: 女",
                                "additionalProperties": false
                              },
                              "phone": {
                                "type": "string",
                                "title": "手机号",
                                "additionalProperties": false
                              },
                              "idCard": {
                                "type": "string",
                                "title": "身份证号",
                                "additionalProperties": false
                              },
                              "isCollege": {
                                "type": "integer",
                                "title": "是否是院校学生",
                                "description": "1: 是 , 0: 否",
                                "format": "int32"
                              },
                              "address": {
                                "type": "string",
                                "title": "联系地址",
                                "additionalProperties": false
                              },
                              "degree": {
                                "type": "integer",
                                "format": "int32",
                                "title": "最高学历",
                                "description": "1:初中, 2:高中, 3:大专, 4:本科, 5:硕士, 6:博士",
                                "additionalProperties": false
                              },
                              "graduateDate": {
                                "type": "string",
                                "format": "date-time",
                                "title": "毕业时间",
                                "additionalProperties": false
                              },
                              "clazzId": {
                                "type": "integer",
                                "format": "int32",
                                "title": "班级ID",
                                "additionalProperties": false
                              },
                              "violationCount": {
                                "type": "integer",
                                "format": "int32",
                                "title": "违纪次数",
                                "additionalProperties": false
                              },
                              "violationScore": {
                                "type": "integer",
                                "format": "int32",
                                "title": "违纪扣分",
                                "additionalProperties": false
                              },
                              "createTime": {
                                "type": "string",
                                "format": "date-time",
                                "title": "创建时间",
                                "additionalProperties": false
                              },
                              "updateTime": {
                                "type": "string",
                                "format": "date-time",
                                "title": "修改时间",
                                "additionalProperties": false
                              },
                              "clazzName": {
                                "type": "string",
                                "title": "班级名称"
                              }
                            },
                            "x-apifox-refs": {},
                            "x-apifox-orders": [
                              "id",
                              "name",
                              "no",
                              "gender",
                              "phone",
                              "idCard",
                              "isCollege",
                              "address",
                              "degree",
                              "graduateDate",
                              "clazzId",
                              "violationCount",
                              "violationScore",
                              "createTime",
                              "updateTime",
                              "clazzName"
                            ],
                            "required": [
                              "idCard",
                              "isCollege",
                              "clazzName"
                            ],
                            "x-apifox-ignore-properties": []
                          },
                          "title": "结果列表",
                          "description": "结果列表"
                        }
                      },
                      "required": [
                        "total",
                        "rows"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410804-run"
      },
      "post": {
        "summary": "添加学生",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "添加学员信息",
        "operationId": "saveUsingPOST_3",
        "tags": [
          "学生管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "title": "姓名"
                  },
                  "no": {
                    "type": "string",
                    "title": "学号"
                  },
                  "gender": {
                    "type": "integer",
                    "format": "int32",
                    "title": "性别",
                    "description": "1: 男 , 2: 女"
                  },
                  "phone": {
                    "type": "string",
                    "title": "手机号"
                  },
                  "idCard": {
                    "type": "string",
                    "title": "身份证号"
                  },
                  "isCollege": {
                    "type": "string",
                    "title": "是否是院校学生",
                    "description": "1: 是 , 0: 否"
                  },
                  "address": {
                    "type": "string",
                    "title": "联系地址"
                  },
                  "degree": {
                    "type": "integer",
                    "format": "int32",
                    "title": "最高学历",
                    "description": "1:初中, 2:高中, 3:大专, 4:本科, 5:硕士, 6:博士"
                  },
                  "graduationDate": {
                    "type": "string",
                    "format": "date-time",
                    "title": "毕业时间"
                  },
                  "clazzId": {
                    "type": "integer",
                    "format": "int32",
                    "title": "班级ID"
                  }
                },
                "x-apifox-refs": {},
                "x-apifox-orders": [
                  "name",
                  "no",
                  "gender",
                  "phone",
                  "idCard",
                  "isCollege",
                  "address",
                  "degree",
                  "graduationDate",
                  "clazzId"
                ],
                "required": [
                  "idCard",
                  "isCollege"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410805-run"
      },
      "put": {
        "summary": "修改学生",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "修改学生信息",
        "operationId": "updateUsingPUT_3",
        "tags": [
          "学生管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410806-run"
      }
    },
    "/students/violation/{id}/{score}": {
      "put": {
        "summary": "违纪处理",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "学生违纪信息处理",
        "operationId": "violationHandleUsingPUT",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "3",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "score",
            "in": "path",
            "description": "score",
            "required": true,
            "example": "5",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410807-run"
      }
    },
    "/students/{ids}": {
      "delete": {
        "summary": "批量删除学生",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "批量删除学生信息",
        "operationId": "deleteUsingDELETE_3",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "path",
            "description": "ids",
            "required": true,
            "example": "3,4,6",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410808-run"
      }
    },
    "/students/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询",
        "operationId": "getInfoUsingGET_2",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410809-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "title": "ID",
            "additionalProperties": false
          },
          "name": {
            "type": "string",
            "title": "姓名",
            "additionalProperties": false
          },
          "no": {
            "type": "string",
            "title": "学号",
            "additionalProperties": false
          },
          "gender": {
            "type": "integer",
            "format": "int32",
            "title": "性别",
            "description": "1: 男 , 2: 女",
            "additionalProperties": false
          },
          "phone": {
            "type": "string",
            "title": "手机号",
            "additionalProperties": false
          },
          "idCard": {
            "type": "string",
            "title": "身份证号",
            "additionalProperties": false
          },
          "isCollege": {
            "type": "integer",
            "title": "是否是院校学生",
            "description": "1: 是 , 0: 否",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "title": "联系地址",
            "additionalProperties": false
          },
          "degree": {
            "type": "integer",
            "format": "int32",
            "title": "最高学历",
            "description": "1:初中, 2:高中, 3:大专, 4:本科, 5:硕士, 6:博士",
            "additionalProperties": false
          },
          "graduationDate": {
            "type": "string",
            "format": "date-time",
            "title": "毕业时间",
            "additionalProperties": false
          },
          "clazzId": {
            "type": "integer",
            "format": "int32",
            "title": "班级ID",
            "additionalProperties": false
          },
          "violationCount": {
            "type": "integer",
            "format": "int32",
            "title": "违纪次数",
            "additionalProperties": false
          },
          "violationScore": {
            "type": "integer",
            "format": "int32",
            "title": "违纪扣分",
            "additionalProperties": false
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "创建时间",
            "additionalProperties": false
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "修改时间",
            "additionalProperties": false
          }
        },
        "title": "Student",
        "x-apifox-orders": [
          "id",
          "name",
          "no",
          "gender",
          "phone",
          "idCard",
          "isCollege",
          "address",
          "degree",
          "graduationDate",
          "clazzId",
          "violationCount",
          "violationScore",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "状态码",
            "description": "1: 成功 , 0: 失败",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "title": "错误提示信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "响应数据",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "Result",
        "x-apifox-orders": [
          "code",
          "msg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "http://test-cn.your-api-server.com",
      "description": "测试环境"
    },
    {
      "url": "http://prod-cn.your-api-server.com",
      "description": "正式环境"
    }
  ]
}